+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0 Unreleased]
+## Added
+- Add tickets frequency: in the fact model you can select how often a ticket is bookable 
+- Rewrited facts list: less code, more simplicity
+- Migrated from turbolinks to hotwire turbo
+- Now you can reserve a ticket from the happening list page
 
 ## [1.0.4] - 2020-09-24
 ### Changed
