+++ app/helpers/happenings_helper.rb
@@ -2,4 +2,22 @@
 
 # this module contain specific helpers for {HappeningsController}
 module HappeningsHelper
+  def happening_title(happening)
+    <<-HTML.html_safe
+      <div class="columns is-mobile is-gapless">
+        <div class="column">
+          <span class="title is-size-5">
+            <i class="fas fa-calendar-day"></i>
+            #{l happening.start_at, format: :short}
+          </span>
+        </div>
+        <div class="column is-narrow has-text-right">
+          <span class="subtitle is-size-6">
+            #{happening.seats_count} / #{happening.max_seats}
+            <i class="fas fa-chair" aria-hidden="true"></i>
+          </span>
+        </div>
+      </div>
+    HTML
+  end
 end
