+++ app/javascript/controllers/form_controller.js
@@ -1,0 +1,15 @@
+import { Controller } from "stimulus"
+import Timeout from 'smart-timeout'
+
+export default class extends Controller {
+  static targets = [ "sender" ]
+
+  sendForm(event) {
+   this.element.requestSubmit();
+  }
+  
+  delayedSendForm(event) {
+    if ( Timeout.exists('textDelay') ) { Timeout.clear('textDelay', true) }
+    Timeout.set('textDelay', () => {this.sendForm(event);}, 750)
+  }
+}
