+++ app/javascript/packs/application.js
@@ -3,23 +3,24 @@
 // a relevant structure within app/javascript and only use these pack files to reference
 // that code so it'll be compiled.
 
-require("@rails/ujs").start()
-require("turbolinks").start()
-require("@rails/activestorage").start()
-require("channels")
+import Rails from "@rails/ujs"
+import "@hotwired/turbo-rails"
+import * as ActiveStorage from "@rails/activestorage"
+import "channels"
 
-
-// Uncomment to copy all static images under ../images to the output folder and reference
-// them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
-// or the `imagePath` JavaScript helper below.
-//
-const images = require.context('../images', true)
-const imagePath = (name) => images(name, true)
+Rails.start()
+ActiveStorage.start()
 
 import "controllers"
 
 require("trix")
 require("@rails/actiontext")
 import './awesome.js'
-import './loader.js'
+
+// Stilesheet
 import '../stylesheets/application'
+require.context("images", true, /\.(png|svg|jpg|gif|ico)$/);
+// const images = require.context('../images', true)
+// const imagePath = (name) => images(name, true)
+
+
