+++ app/javascript/stylesheets/application.sass
@@ -4,7 +4,6 @@
 @import "actiontext"
 @import "trix"
 @import "config"
-@import "~bulma/bulma"
 // @import "~bulma-extensions/bulma-tooltip/src/sass/index"
 // @import "~bulma-extensions/bulma-quickview/src/sass/index"
 // @import "~bulma-switch/src/sass/index"
