+++ app/javascript/stylesheets/footer.sass
@@ -1,32 +1,3 @@
 footer
-  #footer-first
-    padding: 41px 0px
-    .column.menu
-      border-left: $ispra-darker 1px solid
-      p.menu-label
-        font-weight: bold
-        color: $primary
-        font-size: 0.8rem
-        margin-bottom: 4px
-        padding-left: 18px
-      ul.menu-list
-        li
-          span.fa-li
-            font-size: 0.8rem
-            padding-top: 6px
-          a
-            font-size: 0.8rem
-            font-weight: bold
-            color: $grey-dark
-            &:hover
-              background: $ispra-darker
-              color: $white
-  #footer-second
-    background: $primary-darker
-    padding: 41px 0px
-    text-align: center
-    color: $white
-    strong
-      color: $white
-    a
-      color: $white
+  .column.menu
+    border-left: $primary-dark 1px solid
