+++ app/javascript/stylesheets/nav1.sass
@@ -5,5 +5,3 @@
       font-size: 0.9rem
       line-height: 20px
       padding: 0rem 0.75rem
-  .column.logo
-    color: #74a848
