+++ app/javascript/stylesheets/nav2.sass
@@ -1,5 +1,4 @@
 #nav2
-  background: $primary
   padding: 10px
   h1
     margin: 0px
