+++ app/javascript/stylesheets/nav3.sass
@@ -1,15 +1,12 @@
 #nav3
-  background: $ispra-darker 
   .navbar-burger
     color: #ffffff
-    border-left: $rgba-primary-4 solid 1px
+    border-left: $primary solid 1px
   .navbar-brand
-    &.is-editor
-      background: $rgba-secondary-2-4
-    a
+   a
       color: #ffffff
       span
-        color: $rgba-complement-2
+        color: $secondary
   nav.navbar
     background: transparent
     min-height: 40px
