+++ app/views/application/_nav.html.haml
@@ -1,0 +1,22 @@
+#nav3.has-background-primary-dark
+  .container
+    %nav.navbar{role: 'navigation', aria_label: 'main navigation'}
+      .navbar-brand
+        %a.navbar-item.has-text-weight-bold.is-italic.is-size-5{href: root_path, data: {turbo_frame: 'yield'}}
+          %span.icon
+            %i.fas.fa-signature.fa-fw
+          %span P
+          artecipo
+        %a.navbar-burger.burger{role: 'button', aria: {label: 'menu', expanded: 'false'}, data: {action: 'page#toggleMenu'}}
+          <span aria-hidden="true"></span><span aria-hidden="true"></span><span aria-hidden="true"></span>
+      #nav3-menu.navbar-menu{data: {target: 'page.menu'}}
+        .navbar-start
+          = link_to fas_icon('calendar', text: 'Eventi'), root_path, class: "navbar-item", data: {turbo_frame: 'yield'}
+          = link_to fas_icon('ticket-alt', text: 'Prenotazioni'), tickets_path, class: 'navbar-item', data: {turbo_frame: 'yield'}
+        .navbar-end
+          = link_to fas_icon('user', text: current_user.username), tickets_path, class: 'navbar-item'
+          %a.navbar-item{href: destroy_user_session_path, data: {method: :delete}}
+            %span.icon
+              %i.fas.fa-sign-out-alt
+            %span.is-hidden-desktop Disconnettiti
+
