+++ app/views/facts/show.html.haml
@@ -1,4 +1,4 @@
-%section{ data: { controller: 'section', section_url: fact_happenings_path(@fact)}}
+%turbo-frame#yield
   %h3.title.has-text-centered
     %i.fas.fa-calendar
     = @fact.title
@@ -6,17 +6,17 @@
     %aside.column.is-3
       %nav.panel
         %p.panel-heading= t 'site.happening.search'
-        = form_with scope: 'filter', url: fact_happenings_path(@fact), method: :get, data: {remote: true, action: "ajax:success->section#goPage"} do |f|
+        = form_with scope: 'filter', url: fact_happenings_path(@fact), method: :get, data: {turbo_frame: 'happenings', controller: 'form', form_target: 'sender'} do |f|
           .panel-block
             %p.control.has-icons-left
-              = f.text_field :text, placeholder: 'Testo', class: 'input', data: {action: 'keyup->section#delayedSendForm'}
+              = f.text_field :text, placeholder: 'Testo', class: 'input', data: {action: 'keyup->form#delayedSendForm'}
               %span.icon.is-left
                 %i.fas.fa-search
           %label.panel-block
-            = f.radio_button :type, '', selected: true, data: {action: "section#sendForm"}
+            = f.radio_button :type, '', selected: true, data: {action: "form#sendForm"}
             %span= t('future', scope: 'site.happening.scope')
           %label.panel-block
-            = f.radio_button :type, 'history', selected: true, data: {action: "section#sendForm"}
+            = f.radio_button :type, 'history', selected: true, data: {action: "form#sendForm"}
             %span= t('history', scope: 'site.happening.scope')
       %figure.image= image_tag(@fact.image.variant(resize_to_limit: [318,318]).processed) 
     .column
@@ -43,5 +43,4 @@
               %p.heading Prenotazioni
               %p.title.is-size-5= @fact.happenings.sum(:seats_count)
         = @fact.body
-      #happenings{ data: { target: 'section.container' } }
-
+      %turbo-frame#happenings{src: fact_happenings_path(@fact)}
