+++ app/views/happenings/_happening.html.haml
@@ -1,13 +1,4 @@
-%a.box.is-hoverable{href: fact_happening_path(@fact,happening), data: {action: "ajax:success->section#goPage", remote: true}}
-  .columns.is-mobile.is-gapless
-    .column
-      %span.title.is-size-5
-        %i.fas.fa-calendar-day
-        = l happening.start_at, format: :short
-    .column.is-narrow.has-text-right
-      %span.subtitle.is-size-6
-        = happening.seats_count
-        \\
-        = happening.max_seats
-        = fas_icon 'chair'
-  = happening.detail
+= turbo_frame_tag happening do
+  %a.box.is-hoverable.block{href: fact_happening_path(@fact,happening)}
+    = happening_title(happening)
+    = happening.detail
