--- app/views/happenings/register.html.erb
@@ -1,2 +1,0 @@
-<h1>Happenings#register</h1>
-<p>Find me in app/views/happenings/register.html.erb</p>
