+++ app/views/tickets/_remove.html.haml
@@ -1,0 +1,14 @@
+.notification.is-success.is-light.has-text-centered N. <b>#{ticket.seats}</b> posti prenotati #{l ticket.updated_at, format: :detailed} da <b>#{ticket.user.username}</b>
+= form_with model: [fact, happening, ticket], method: :delete do |f|
+  %fieldset
+    .field.has-addons.has-icons-left
+      .control
+        = f.label :seats, class: 'button', disabled: true do
+          = fas_icon 'chair', text: 'Posti' 
+      .control.is-expanded
+        .select.is-fullwidth= f.select :seats, [ticket.seats], {}, disabled: true
+        %p.help.is-danger= @ticket.errors.messages[:seats].join(' - ')
+      .control
+        = f.button fas_icon('trash', text: 'Annulla prenotazione'), class: 'button is-warning', data: {confirm: 'Attenzione: vuoi veramente annullare la prenotazione?'}, disabled: !ticket.saleable?
+- if @ticket.start_at < Time.zone.now
+  .notification.is-success.is-warning.has-text-centered Prenotazioni terminate #{l @happening.stop_sale_at, format: :detailed}
