+++ app/views/tickets/_ticket.html.haml
@@ -1,4 +1,4 @@
-%a.box{href: fact_path(ticket.happening.fact), data: {action: "ajax:success->page#goPage", remote: true}}
+%a.box{href: fact_path(ticket.happening.fact), data: {turbo_frame: 'yield'}}
   .columns.is-mobile.is-gapless
     .column.is-narrow.is
       .figure.image.is-boxed-left-image= image_tag ticket.happening.fact.image.variant(resize_to_limit: [150, 68]).processed
