--- app/views/tickets/create.html.erb
@@ -1,2 +1,0 @@
-<h1>Tickets#create</h1>
-<p>Find me in app/views/tickets/create.html.erb</p>
