--- app/views/tickets/destroy.html.erb
@@ -1,2 +1,0 @@
-<h1>Tickets#destroy</h1>
-<p>Find me in app/views/tickets/destroy.html.erb</p>
