+++ app/views/tickets/form_remove.html.haml
@@ -1,0 +1,40 @@
+%turbo-frame#happenings
+  %h3.title.is-size-4.has-text-centered= fas_icon 'calendar-day', text: 'Prenotazione'
+  .box
+    .columns.is-mobile.is-gapless
+      .column
+        %span.title.is-size-5
+          = fas_icon 'calendar-day'
+          = l @happening.start_at, format: :short
+      .column.is-narrow.has-text-right
+        %span.subtitle.is-size-6
+          = @happening.seats_count
+          \\
+          = @happening.max_seats
+          = fas_icon 'chair'
+    - if @ticket.persisted?
+      .notification.is-success.is-light.has-text-centered N. <b>#{@ticket.seats}</b> posti prenotati #{l @ticket.updated_at, format: :detailed} da <b>#{@ticket.user.username}</b>
+    - if @ticket.saleable?
+      = form_with model: @ticket, url: fact_happening_tickets_path(@fact,@happening), method: :post do |f|
+        %fieldset
+          .field.has-addons.has-icons-left
+            .control
+              = f.label :seats, class: 'button', disabled: @ticket.persisted? do
+                = fas_icon 'chair', text: 'Posti' 
+            .control.is-expanded
+              .select.is-fullwidth= f.select :seats, (1..@happening.max_seats_for_ticket), {}, disabled: @ticket.persisted?
+              %p.help.is-danger= @ticket.errors.messages[:seats].join(' - ')
+            .control
+              - if @ticket.persisted?
+                = link_to fas_icon('trash', text: 'Annulla prenotazione'), fact_happening_ticket_path(@fact, @happening, @ticket), method: :delete, class: 'button is-warning', data: {confirm: 'Attenzione: vuoi veramente annullare la prenotazione?', turbo_frame: 'happenings'}
+              -else
+                = f.button fas_icon('ticket-alt', text: 'Prenota'), class: 'button is-success'
+    - if @ticket.start_at > Time.zone.now
+      .has-text-centered Prenotazioni aperte dalle #{l @happening.start_sale_at, format: :time_first} alle #{l @happening.stop_sale_at, format: :time_first}
+    - else
+      .notification.is-success.is-warning.has-text-centered Prenotazioni terminate #{l @happening.stop_sale_at, format: :detailed}
+  .has-text-centered
+    = link_to fas_icon('chevron-left', text: 'Indietro'), fact_happenings_path(@fact), class: 'button is-light', data: {action: "ajax:success->section#goPage", remote: true}
+    = link_to fas_icon('ticket-alt', text: 'Prenotazioni'),  tickets_path, class: 'button is-light', data: {action: "ajax:success->page#goPage", remote: true}
+
+
