+++ app/views/tickets/list.html.haml
@@ -1,4 +1,5 @@
-#tickets= render @tickets
-%hr
-= pagy_bulma_nav(@pagy).html_safe
+%turbo-frame#tickets
+  #tickets= render @tickets
+  %hr
+  = pagy_bulma_nav(@pagy).html_safe
