+++ package.json
@@ -6,6 +6,7 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-brands-svg-icons": "^5.13.0",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@hotwired/turbo-rails": "^7.0.0-beta.5",
     "@rails/actioncable": "^6.0.0",
     "@rails/actiontext": "^6.0.2-2",
     "@rails/activestorage": "^6.0.0",
@@ -16,8 +17,7 @@
     "slim-select": "^1.26.0",
     "smart-timeout": "^2.5.0",
     "stimulus": "^1.1.1",
-    "trix": "^1.2.3",
-    "turbolinks": "^5.2.0"
+    "trix": "^1.2.3"
   },
   "version": "0.1.0",
   "devDependencies": {
