+++ public/footer.html
@@ -1,69 +1,66 @@
+<turbo-frame id="footer">
+<footer>
 
-<div id="footer-first" class="has-background-grey-lighter">
+
+<div id="footer-first" class="has-background-grey-lighter py-6 has-text-weight-bold">
 <div class="container">
-<div class="columns is-desktop">
+<div class="columns has-4-desktop has-2-tablet">
 <div class="column menu">
-<p class="menu-label">Intranet</p>
-<ul class="menu-list fa-ul">
+<p class="menu-label has-text-primary-dark">Intranet</p>
+<ul class="menu-list fa-ul is-size-7">
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Home Page</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>
+Home Page</a>
 </li>
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Rubrica</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Rubrica</a>
 </li>
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Cloud</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Cloud</a>
 </li>
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Struttura</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Struttura</a>
 </li>
 
 </ul>
 </div>
 <div class="column menu">
-<p class="menu-label">Assistenza</p>
-<ul class="menu-list fa-ul">
+  <p class="menu-label has-text-primary-dark">Assistenza</p>
+<ul class="menu-list fa-ul is-size-7">
 
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Giri</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Giri</a>
 </li>
 </ul>
-<p class="menu-label">Posta elettronica</p>
-<ul class="menu-list fa-ul">
+<p class="menu-label has-text-primary-dark">Posta elettronica</p>
+<ul class="menu-list fa-ul is-size-7">
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Web Mail</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Web Mail</a>
 </li>
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Configurazione</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Configurazione</a>
 </li>
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Avvisi</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Avvisi</a>
 </li>
 </ul>
 </div>
 <div class="column menu">
-<p class="menu-label">Applicazioni</p>
-<ul class="menu-list fa-ul">
+  <p class="menu-label has-text-primary-dark">Applicazioni</p>
+<ul class="menu-list fa-ul is-size-7">
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Magazzino</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Magazzino</a>
 </li>
 <li>
-  <span class="fa-li "><i class="fas fa-link"></i></span>
-  <a target="_blank" href="/">Segreteria medica</a>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Sale riunioni</a>
+</li>
+<li>
+  <a target="_blank" href=""><span class="fa-li "><i class="fas fa-link"></i></span>Segreteria medica</a>
 </li>
 </ul>
 </div>
 <div class="column menu">
-<p class="menu-label">Portale</p>
+  <p class="menu-label has-text-primary-dark">Portale</p>
 <ul class="menu-list fa-ul">
 
 </ul>
@@ -71,9 +68,8 @@
 </div>
 </div>
 </div>
-<div id="footer-second">
-<div class="container">
-  Partecipo | https://github.com/isprambiente/Partecipo
+<div class="py-6 has-text-centered has-background-primary-dark has-text-white">
+  Copyright 2020 <span class='has-text-weight-bold'>ISPRA</span> Istituto Superiore per la Protezione e la Ricerca Ambientale
 </div>
-</div>
-
+</footer>
+</turbo-frame>
