+++ public/head.html
@@ -1,25 +1,25 @@
-  <div id="nav1">
-    <div class="container">
+<turbo-frame id='nav1'>
+    <div class="container is-hidden-mobile">
       <nav aria_label="main navigation" class="navbar" role="navigation">
         <div class="navbar-menu">
           <div class="navbar-end">
-            <a class="navbar-item" target="_blank" href="/">Group</a>
-            <a class="navbar-item" target="_blank" href="/">link</a>
-            <a class="navbar-item" target="_blank" href="/">other</a>
+            <a class="navbar-item" target="_blank" href="http://www.isprambiente.gov.it">ISPRA</a>
+            <a class="navbar-item" target="_blank" href="https://intranet.isprambiente.it">Intranet</a>
+            <a class="navbar-item" target="_blank" href="https://rubrica.intranet.isprambiente.it">Rubrica</a>
           </div>
         </div>
       </nav>
     </div>
   </div>
-  <div id="nav2">
+  <div id="nav2" class="has-background-primary">
     <div class="container">
       <div class="columns is-mobile is-gapless">
-        <h1>
-          <a target="_blank" href="/"><i class="fas fa-signature fa-fw fa-3x"></i></a>
-        </h1> 
+        <div class="column logo">
+	  <h1><a target="_blank" href="/"><i class="fas fa-signature fa-fw fa-3x"></i></a></h1>  
+        </div>
         <div class="column text is-hidden-mobile">
           <h1 class="has-text-right"><a href="/">An Example Title<br> For An Example Group</a></h1>
         </div>
       </div>
     </div>
-  </div>
+</turbo-frame>
